Make debugger list and debugger frame independent with no circular references.
